Test table renders correct data
setup test for table and extracted code for column from ERC20Balance so it can be reused
