+++ package.json
@@ -8,9 +8,6 @@
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@lifi/sdk": "^1.6.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
     "@typescript-eslint/parser": "^5.40.1",
     "@walletconnect/web3-provider": "^1.6.6",
     "@web3auth/web3auth": "^1.1.4",
@@ -34,7 +31,7 @@
   "scripts": {
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
-    "test": "node scripts/test.js",
+    "test": "react-scripts test",
     "eject": "react-app-rewired eject",
     "devchain": "node Truffle/scripts/devChain.js",
     "connect": "moralis-admin-cli connect-local-devchain",
@@ -68,7 +65,10 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^28.1.4",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "12.1.5",
+    "@testing-library/user-event": "^14.2.0",
+    "@types/jest": "^29.2.3",
     "@types/node": "^18.0.0",
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
@@ -91,6 +91,7 @@
     "react-app-rewired": "^2.2.1",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
+    "ts-jest": "^29.0.3",
     "typescript": "^4.7.4",
     "url": "^0.11.0"
   }
